Add initial path for primary branch, fix font pitch to 0.5 of height
The primary branch (`master` or whichever other branch is the primary) has now
an automatic merge path that leads to the beginning of time, much like branches
that have been created before the first commit.

The font pitch is also fixed to half the height, as is usually the case with
monospace fonts.